Linux compatibility

+++ libs/RakNet/BitStream.h
@@ -1537,7 +1537,7 @@ namespace RakNet
 			y=0.0;
 		else
 		{
-			ReadCompressed((float)cy);
+			ReadCompressed(cy);
 			y=cy;
 			//Read(sy);
 			//y=((float)sy / 32767.5f - 1.0f);
@@ -1554,7 +1554,7 @@ namespace RakNet
 			//		return false;
 
 			//	z=((float)sz / 32767.5f - 1.0f);
-			if (!ReadCompressed((float)cz))
+			if (!ReadCompressed(cz))
 				return false;
 			z=cz;
 		}
